Refactor redundant page fetch in dashboard
