+++ react-app/src/components/Dashboard/Dashboard.js
@@ -1,4 +1,6 @@
-import React from "react";
+import React, { useEffect } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { TH_getUserPages } from "../../store/page";
 import BookmarkedPages from "./BookmarkedPages";
 
 import "./Dashboard.css";
@@ -6,14 +8,21 @@ import RecentPages from "./RecentPages";
 import Scratchpad from "./Scratchpad";
 
 const Dashboard = () => {
+    const dispatch = useDispatch();
+
+    const userPages = useSelector(state => state.pages.user);
+
+    useEffect(() => {
+        dispatch(TH_getUserPages());
+    }, [dispatch]);
 
     return (
         <div id="dashboard-wrapper">
             <div id="dashboard-bookmarks-wrapper">
-                <BookmarkedPages />
+                <BookmarkedPages userPages={userPages} />
             </div>
             <div id="dashboard-pages-wrapper">
-                <RecentPages />
+                <RecentPages userPages={userPages} />
             </div>
             <div id="dashboard-scratchpad-wrapper">
                 <Scratchpad />
